abab

+++ Ev-System/src/main/java/com/example/Ev/System/dto/WorkLogDto.java
@@ -3,11 +3,13 @@ package com.example.Ev.System.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class WorkLogDto {
     private Integer staffId;
     private Integer appointmentId;
-    private Double hoursSpent;
+    private BigDecimal hoursSpent;
     private String tasksDone;
 }
